Split several of the larger WML tags into their own file
